requirements.txt
aiogram==3.15.0
pyrogram==2.0.106
tgcrypto==1.2.5
supabase==2.11.0
python-dotenv==1.0.1
structlog==24.4.0
